Backend/requirements.txt
flask==3.0.0
openai==1.59.7
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
flask-sock==0.7.0
websocket-client==1.7.0
numpy==1.26.3
pydub==0.25.1
soundfile==0.12.1 